load_generator/requirements.txt
# load_generator/requirements.txt
kafka-python==2.0.2
pandas==1.5.3
numpy==1.24.3